sviluppo registra vendita in negozio

+++ src/main/java/it/unicam/pawm/c3/Negozio.java
@@ -12,7 +12,6 @@ import org.hibernate.annotations.LazyCollectionOption;
 
 import javax.persistence.*;
 import java.util.ArrayList;
-import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
 
@@ -23,7 +22,7 @@ public class Negozio {
     @Id
     @Column(name="id")
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private Long id;
     private String nome;
     private String indirizzo;
     private String p_iva;
@@ -112,11 +111,11 @@ public class Negozio {
         this.p_iva = p_iva;
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
